requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
google-auth
gcsfs
numpy
pandas
pyarrow
tensorflow-cpu; python_version < "3.12"
